+++ hw/xnest/Display.h
@@ -33,7 +33,7 @@ extern Drawable xnestDefaultDrawables[MAXDEPTH + 1];
 extern Pixmap xnestIconBitmap;
 extern Pixmap xnestScreenSaverPixmap;
 extern uint32_t xnestBitmapGC;
-extern unsigned long xnestEventMask;
+extern uint32_t xnestEventMask;
 
 void xnestOpenDisplay(int argc, char *argv[]);
 void xnestCloseDisplay(void);
